Bullet Shell - Managed the Game (and managed to break it)
note: this is kind of an emergency push to save all of my scripting progress. Unfortunately, I neglected to save the changes I made to the scene itself, but that's okay, since I can easily re-do it when I get home. What happened was I absent-mindedly put a while loop (based on user input) inside of Update(), so the game stopped incrementing frames -- and as a result, also prevented itself from receiving new user input, which meant that the while loop's condition never returned false. Since this was all on a campus computer, the Windows Task Manager was disabled, so I now have no way to close Unity except by restarting the computer -- and if the computer re-images itself while restarting, my progress would be lost. And so, I'm pushing my scripts now, and I'll just fix the scene tomorrow at home.

Anyways, the update log:
 - Created a GameManager script, and moved various gameplay functionalities into it from the other scripts.
    + Unfortunately, I froze Unity before I could do a full playtest, so I'm not 100% sure if it worked properly. However, from what I saw, everything (aside from the literally game-breaking bug) worked fine. I'll test it further at home.
    + List of scripts affected:
          - LifeScoreManager: all code inside this was transferred to GameManager, and once testing is complete, this script will be deleted.
          - PlayManager: See above.
          - CollisionDetection: Transferred the game over code to GameManager.
          - PlayerController: transferred cursor toggling code to GameManager.
 - Created a Main Menu (didn't save objects, but all the code is still okay)
 - Created a playTime variable to use instead of Time.time. It's probably less precise, but it resets itself each time the scene loads, and it only starts counting when the round actually starts.
 - Fixed the colors for the Life pickup and the Speed pickup.
 - Added background music (also didn't save).
 - Found sound effects to use, but didn't get to add them yet. I'm just going to put them separately in the repository until I can import them to the project when I work on it at home.

To-Do:
 - restore lost progress to the scene
 - stop player from moving around before the game starts.
 - add visual feedback for when the player is hiding in their shell.
 - add sound effects for when powerups are collected and when bullets hit the shell while the player is hiding.
 - (optional) change isGameOver to isGameRunning, for readability.
 - (optional) one last time, refactor the code for the patterns/bullets for readability and (maybe) efficiency. Store the graph equations as functions inside of the PatternController script, then simply have the bullets call those functions. This would cut down the number of variables the scripts would need to track, making it much more readable, and possibly more efficient.